Generalize interactive mode implementation for Llama and Mistral
