+++ src/main/java/com/example/auxiliary/format/LlamaChatFormat.java
@@ -21,7 +21,7 @@ public class LlamaChatFormat implements ChatFormat {
 
     public LlamaChatFormat(LlamaTokenizer tokenizer) {
         this.tokenizer = tokenizer;
-        Map<String, Integer> specialTokens = this.tokenizer.getSpecialTokens();
+        Map<String, Integer> specialTokens = tokenizer.getSpecialTokens();
         this.beginOfText = specialTokens.get("<|begin_of_text|>");
         this.startHeader = specialTokens.get("<|start_header_id|>");
         this.endHeader = specialTokens.get("<|end_header_id|>");
@@ -31,60 +31,40 @@ public class LlamaChatFormat implements ChatFormat {
         this.stopTokens = Set.of(endOfText, endOfTurn);
     }
 
-    public Tokenizer getTokenizer() {
-        return tokenizer;
-    }
+    @Override
+    public int getBeginOfText() { return beginOfText; }
 
-    public Set<Integer> getStopTokens() {
-        return stopTokens;
-    }
+    @Override
+    public Set<Integer> getStopTokens() { return stopTokens; }
 
-    public List<Integer> encodeHeader(LlamaChatFormat.Message message) {
+    @Override
+    public List<Integer> encodeHeader(Message message) {
         List<Integer> tokens = new ArrayList<>();
-        LlamaTokenizer llamaTokenizer = (LlamaTokenizer) this.tokenizer;
         tokens.add(startHeader);
-        tokens.addAll(llamaTokenizer.encodeAsList(message.role().name()));
+        tokens.addAll(tokenizer.encodeAsList(message.role().name()));
         tokens.add(endHeader);
-        tokens.addAll(llamaTokenizer.encodeAsList("\n"));
-        return tokens;
-    }
-
-    public List<Integer> encodeMessage(LlamaChatFormat.Message message) {
-        List<Integer> tokens = this.encodeHeader(message);
-        tokens.addAll(this.tokenizer.encodeAsList(message.content().strip()));
-        tokens.add(endOfTurn);
-        return tokens;
-    }
-
-    public List<Integer> encodeDialogPrompt(boolean appendAssistantTurn, List<LlamaChatFormat.Message> dialog) {
-        List<Integer> tokens = new ArrayList<>();
-        tokens.add(beginOfText);
-        for (LlamaChatFormat.Message message : dialog) {
-            tokens.addAll(this.encodeMessage(message));
-        }
-        if (appendAssistantTurn) {
-            // Add the start of an assistant message for the model to complete.
-            tokens.addAll(this.encodeHeader(new LlamaChatFormat.Message(LlamaChatFormat.Role.ASSISTANT, "")));
-        }
+        tokens.addAll(tokenizer.encodeAsList("\n"));
         return tokens;
     }
 
     @Override
-    public int getBeginOfText() {
-        return beginOfText;
+    public List<Integer> encodeMessage(Message message) {
+        List<Integer> tokens = encodeHeader(message);
+        tokens.addAll(tokenizer.encodeAsList(message.content().strip()));
+        tokens.add(endOfTurn);
+        return tokens;
     }
 
-    public record Message(LlamaChatFormat.Role role, String content) {
-    }
-
-    public record Role(String name) {
-        public static LlamaChatFormat.Role SYSTEM = new LlamaChatFormat.Role("system");
-        public static LlamaChatFormat.Role USER = new LlamaChatFormat.Role("user");
-        public static LlamaChatFormat.Role ASSISTANT = new LlamaChatFormat.Role("assistant");
-
-        @Override
-        public String toString() {
-            return name;
+    public List<Integer> encodeDialogPrompt(boolean appendAssistantTurn, List<Message> dialog) {
+        List<Integer> tokens = new ArrayList<>();
+        tokens.add(beginOfText);
+        for (LlamaChatFormat.Message message : dialog) {
+            tokens.addAll(encodeMessage(message));
         }
+        if (appendAssistantTurn) {
+            // Add the start of an assistant message for the model to complete.
+            tokens.addAll(encodeHeader(new Message(ChatFormat.Role.ASSISTANT, "")));
+        }
+        return tokens;
     }
 }
