+++ src/main/java/com/example/tokenizer/impl/Tokenizer.java
@@ -9,6 +9,14 @@ public interface Tokenizer {
     String                  regexPattern();
     Map<String, Integer>    getSpecialTokens();
     boolean                 isSpecialToken(int tokenIndex);
+    /**
+     * Determines if a token should be displayed during streaming output.
+     * This filters out special tokens, control characters, or other non-displayable content.
+     *
+     * @param token the token to check
+     * @return true if the token should be displayed to the user, false otherwise
+     */
+    boolean                 shouldDisplayToken(int token);
     List<Integer>           encode(String text, Set<String> allowedSpecial);
     List<Integer>           encodeAsList(String text);
     String                  decode(List<Integer> tokens);
